+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Fixed
 
 * Typescript fixes
+* add shape as second argument for `sceneFunc` and `hitFunc`
 
 ## [2.1.4][2018-06-15]